Add three overrides and bring code into line
`function-paren-newline` and `no-confusing-arrow` both get disabled here, we meaning don't need parens around JSX or around the ternaries in arrow functions. Also, `comma-dangle` is added with a special option which enables us to have our JSX blocks always properly aligned with no trailing comma and a closing paren on a separate line.

Changes to code are:

* Brackets removed in numerous places around JSX blocks where not required
* When JSX lines are very long or when components are mounted with children they generally begin on their own new line
* No trailing commas after JSX and closing parens on next line
* Fixes to tabbing of children
* `max-len` seems not to be taken into account by ESLint, but I have moved children onto a new line when line length would exceed 100 chars and generally kept to this rule elsewhere too
* No blank line between packages which are core ones, but kept one between core ones and our ones, and between our ones and Storybook ones